Support system installs of LMDB (ticket #40).
This requires a little messing to work reliably with cffi, since cffi
hashes the state of the compile environment at runtime, we need to
ensure setup.py state is preserved so that any existing cffi cached
build remains valid.

To do that we capture the necessary chunks of configuration present
during setup.py and persist them into a lmdb/_config.py module.