Touch pages of MDB_val with GIL released. (#65, #86)
This seems to be the simplest, lowest overhead and most portable
solution to the GIL page fault problem: simply loop over the MDB_val,
touching one byte every 4kb. Regardless of whether the value lives in a
leaf page or an overflow page, is 100 bytes or 64kb, this should ensure
on all but the most heavily overloaded machines that the pages are in
RAM before PyString_FromStringAndSize() attempts to copy them.