Fix transaction being removed from reader table when deallocated in fork
If you have the following code:

    txn = env.begin()
    if os.fork() == 0: # child
        return
    os.wait()

Then when the fork exits we will deallocate the transaction which
cleared it from the reader table. This meant that the still live
transaction in the parent wasn't tracked and so a writer was free to
remove the pages it was using.

We now don't clear the transaction if we're in a sub-process. This
doesn't guard against issues _using_ the transaction in the child
process but avoids a previously unavoidable issue using it in the parent
while allowing forking.