Prevent some accidental use of LMDB objects by child process (#375)
Default max_spare_txns on cpython impl to 0.

Also, don't create spare transactions if the above parameter is true.

This prevents hanging open transactions across forks. This was exhibited
by a MDB_BAD_RSLOT error.

Fixes #346.

Grab most of the changes suggested in #363. Thank you to callumwalker for 
their contribution. This records the PID at environment creation time and
prevents deletion of transactions when we're not on that PID.